Moved the development tag to the sidebar folder.
